Set token expiry for CF CLI UAA client
We want to force login sessions to last no longer than a day, due to a
recommendation in a recent penetration test. The access token validity
is set via cf-deployment, but we set the refresh token validity via our
own ops files, so we add a test to make sure the value is set correctly
and cf-deployment does not override our value.